+++ src/subcommands/mod.rs
@@ -1,6 +1,5 @@
 pub mod list;
 mod migrate;
-pub mod modpack;
 pub mod profile;
 mod remove;
 mod upgrade;